refactor(all): 폴더 정리
