--- src/main/kotlin/CardPackGenerator.kt
@@ -1,5 +1,0 @@
-class CardPackGenerator {
-    fun createCards(): Cards {
-        return Cards(buildList { Rank.values().forEach { rank -> addAll(Suit.values().map { suit -> Card(rank, suit) }) } })
-    }
-}
