+++ src/main/kotlin/domain/CardPicker.kt
@@ -1,3 +1,8 @@
+package domain
+
+import model.Card
+import model.Cards
+
 class CardPicker(private val cards: Cards) {
     private var cursor = 0
     fun pick(): Card {
