+++ src/main/kotlin/model/Card.kt
@@ -1,1 +1,3 @@
+package model
+
 data class Card(val rank: Rank, val suit: Suit)
