+++ src/main/kotlin/model/Dealer.kt
@@ -1,3 +1,5 @@
+package model
+
 class Dealer(cards: Cards) : Participant(cards) {
     override fun isHit(): Boolean {
         if (cards.sum() < 17) return true
