+++ src/main/kotlin/model/Participant.kt
@@ -1,3 +1,5 @@
+package model
+
 abstract class Participant(val cards: Cards) {
     abstract fun isHit(): Boolean
     fun isBurst(): Boolean {
