+++ src/main/kotlin/model/Player.kt
@@ -1,3 +1,5 @@
+package model
+
 class Player(val name: Name, cards: Cards) : Participant(cards) {
     override fun isHit(): Boolean = !isBurst()
 }
