+++ src/main/kotlin/model/Suit.kt
@@ -1,3 +1,5 @@
+package model
+
 enum class Suit {
     DIAMOND, CLOVER, SPACE, HEART
 }
