+++ src/test/kotlin/blackjack/domain/CardPackGeneratorTest.kt
@@ -1,6 +1,6 @@
-package blackjack
+package blackjack.domain
 
-import CardPackGenerator
+import domain.CardPackGenerator
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
