+++ src/test/kotlin/blackjack/model/CardsTest.kt
@@ -1,7 +1,9 @@
-package blackjack
+package blackjack.model
 
-import Card
-import Cards
+import model.Card
+import model.Cards
+import model.Rank
+import model.Suit
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
