+++ src/test/kotlin/blackjack/model/GameResultTest.kt
@@ -1,10 +1,13 @@
-package blackjack
+package blackjack.model
 
-import Card
-import Cards
-import Dealer
-import Name
-import Player
+import model.Card
+import model.Cards
+import model.Dealer
+import model.GameResult
+import model.Name
+import model.Player
+import model.Rank
+import model.Suit
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 
@@ -39,7 +42,17 @@ class GameResultTest {
     fun `딜러 20점 1승1패 플레이어 19점 패배, 21점 승리`() {
         val dealer = Dealer(Cards(listOf(Card(Rank.JACK, Suit.HEART), Card(Rank.JACK, Suit.DIAMOND))))
         val player1 = Player(Name("jason"), Cards(listOf(Card(Rank.JACK, Suit.HEART), Card(Rank.NINE, Suit.DIAMOND))))
-        val player2 = Player(Name("pobi"), Cards(listOf(Card(Rank.JACK, Suit.SPACE), Card(Rank.KING, Suit.DIAMOND), Card(Rank.ACE, Suit.DIAMOND))))
+        val player2 = Player(
+            Name("pobi"),
+            Cards(
+                listOf(
+                    Card(Rank.JACK, Suit.SPACE), Card(Rank.KING, Suit.DIAMOND),
+                    Card(
+                        Rank.ACE, Suit.DIAMOND
+                    )
+                )
+            )
+        )
         val result = GameResult.of(dealer, listOf(player1, player2))
 
         assertThat(result.getDealerWinResult()).isEqualTo(1)
